requirement.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.20
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
importlib-resources==5.4.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.1.1
tzdata==2023.3
zipp==3.6.0
